+++ docs/include/samplecode/SoundTouchClient/GetWirelessProfile.py
@@ -11,16 +11,16 @@ try:
     client:SoundTouchClient = SoundTouchClient(device)
 
     # get real-time configuration from the device.
-    wirelessProfile:WirelessProfile = client.GetWirelessProfile()
+    wirelessProfile:WirelessProfileActive = client.GetWirelessProfile()
     print(wirelessProfile.ToString())
 
     # get cached configuration, refreshing from device if needed.
-    wirelessProfile:WirelessProfile = client.GetWirelessProfile(False)
+    wirelessProfile:WirelessProfileActive = client.GetWirelessProfile(False)
     print("\nCached configuration:\n%s" % wirelessProfile.ToString())
 
     # get cached configuration directly from the configuration manager dictionary.
     if SoundTouchNodes.getActiveWirelessProfile.Path in client.ConfigurationCache:
-        wirelessProfile:WirelessProfile = client.ConfigurationCache[SoundTouchNodes.getActiveWirelessProfile.Path]
+        wirelessProfile:WirelessProfileActive = client.ConfigurationCache[SoundTouchNodes.getActiveWirelessProfile.Path]
         print("\nCached configuration, direct:\n%s" % wirelessProfile.ToString())
         
 except Exception as ex: