+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.83 ] - 2025/11/02
+
+  * Added method `SoundTouchClient.AddWirelessProfile` to add a new wireless profile to the target device network configuration.
+  * Added method `SoundTouchClient.SetLanguage` to set a new device language code.
+
 ###### [ 1.0.82 ] - 2025/07/27
 
   * Updated package requirement from `smartinspectpython>=3.0.37` to `smartinspectpython>=3.0.38`.